requirements.txt
fastapi
uvicorn
uvicorn[standard]  # Included for Gunicorn Uvicorn Worker support
pydantic
google-genai       # Official SDK for Gemini
gunicorn   
python-dotenv         # Production-ready web server